Hardcode module version in git2go binary name variable
git2go package is dynamically adding the module version to the
BinaryName variable. Only production build contains the module
version though, leading the test builds to have the binary name as
'gitaly-git2go-'. This makes it more difficult to centralize the
binary location resolving as we'd have to differentiate this case.
For now, let's just hard code the v15 suffix to the variable. We
can soon drop the suffix entirely when the backwards compatibility
concerns are dealt with by pinning the auxiliary binaries a given
Gitaly can invoke.