Don't build packed binaries as part of prepare-tests in CI
Our makefile forcefully rebuilds Gitaly binaries on every run. This
causes the prepare-tests target to always rebuild the packed binaries.
While the packed executables are a dependency that should be built prior
to running tests, they can't be built in CI due to the CI running
without privileges. Skip rebuilding the packed exectutables as part of
prepare-tests if we are running in CI. SKIP_RSPEC_BUILD is used for the
same reason to skip rspec build. This commit renames the variable to a
more general name and skips the packed binary building in CI during
tests.