gittest: Fix the default path for created repositories
If not specified, the `CreateRepo()` test helper falls back to compute a
default repository path for our hashed storage. This default path is
wrong though: it's using `newDiskHash()` directly and is thus missing
the `@hashed` prefix.

Fix this issue by using `NewRepositoryName()` instead. While this should
not really impact anything, it's still sensible to do the right thing
and match our production setup.