featureflag: Add accessor to access list of feature flag definitions
The list of available feature flags is currently a globally accessible
variable that can be changed at will by other packages. Let's change
this to instead use an accessor function `DefinedFlags()`: this forbids
any change to the feature flags and allows us to iterate on the way we
store flags.