ci: Test with Go v1.18
We have recently started to require Go 1.17 and later. Interestingly
enough, our CI has not been updated to use that new version yet. Because
of whatever reason this seems to be fine, and Go doesn't complain about
it when running with v1.16.

Let's fix this regardless and start testing with Go v1.17 and v1.18,
only.