git: Fix missing consistency checks for clones
There are three different cases where we clone repositories:

    - When creating an object pool from a preexisting local repository.

    - When creating a fork from a preexisting local repository.

    - When creating a repository from an external URL.

We don't currently run object consistency checks in either of those
cases, which means that we don't detect objects that are a known attack
vectors for certain exploits. Consequentially, we can easily end up as a
carrier for any such broken objects and help distributing them.

Now the first two cases are special: we already have the object data on
disk anyway, so if we started to complain about any such preexisting
object when creating an object pool or fork then it would become
impossible to do so for already existing repositories. In the past our
stance has been that we don't accept _new_ broken objects, but that we
continue to accept repositories that had such baggage due to legacy
reasons. So we don't change the current behaviour for those two cases
and keep on accepting broken objects.

The third case is different though: we pull in objects from an external
repository that may easily be controlled by an adversary. It would thus
be easy to pull in objects that are known to exploit vulnerabilities in
Git. And given our stance that we don't accept new known-broken objects
in repositories it makes perfect sense to start checking objects pulled
in via CreateRepositoryFromURL.

Let's globally enable object consistency checks for all invocations of
git-clone(1). While we only want to enable them for the third case
anyway, this is about defense in depth by doing the safe thing by
default. We fix creation of forks by explicitly disabling the checks for
its invocation of git-clone(1). And for creation of object pools we
don't have to do anything as they use the local file transport of Git,
which only performs a copy of the repository without performing any
consistency checks whatsoever.

Note that this breaks any cases where customers whish to import external
projects that may have broken objects for legacy reasons. While this is
very much intended, this may also break cases where consistency checks
pull in objects that are broken in a "benign" way. For any such benign
brokenness we have in the past created excemptions to again allow such
objects to unbreak valid usecases as they come up.

On the other hand, we do not accept excemptions for consistency checks
which test for known exploits. These are not and will never be accepted
by default. Self-hosted customers are free to override these with their
own Gitaly configuration by adding excemptions via `[[git.config]]`
sections.

Changelog: security