go: Update gitlab-shell dependency to v14.8.0
Upgrade the gitlab-shell dependency to v14.8.0. This upgrade brings
with it an incompatible change to the way the `GitlabNetClient` is
constructed, which doesn't accept the `SelfSigned` parameter anymore.

Upstream has already gone through the deprecation phase of this
parameter in the configuration, so in theory nobody should rely on it
anymore. So consequentially, we should be fine to just drop it without
any replacement ourselves.