helper: Inline input and expectations for `Errorf()` tests
The tests for `Errorf()`-style error wrapping are quite hard to
understand because of the different input and output variables. It may
just be me, but I think it's hard to understand what the difference
between `input`, `inputGRPC`, `inputGRPCFmt` and `inputGRPCCode` is at a
detailed level.

Let's inline these variables to make the testcases mostly self-contained
and thus easier to understand.