Extract runtime directory creation logic into a method
The logic to create the runtime directory currently lies in
setDefaults that is called to set up the default values of the config.
We'll need more control over this soon as we want to prune the old
runtime directories on start up. It also allows us to move the
directory creation outside of the config loading, given it's better
if one can load a config without creating directories as a side effect.