Colocate internal socket dir creation with runtime dir creation
The internal socket directory is expected to be present in the
runtime directory. The config loading process is currently creating
the directory when setting defaults. As we are about to move the
runtime directory creation outside of the config loading process,
let's colocate the internal socket directory creation with the runtime
directory creation as it depends on it. Along with this logic, the
validation logic is colocated with the runtime directory creation
logic as it's not possible to validate the internal socket before the
runtime directory is setup.

Some test cases have been removed as they no longer make sense, such
as the non-existent internal socket dir tests. SetupRuntimeDirectory
always creates the directory.