Setup runtime directory after config has been loaded
Initially, the RuntimeDir contains the parent directory where all of
the runtime directories should be created in. The configuration is then
loaded, and the RuntimeDir is overwritten with the actual created runtime
directory. In order to prune the runtime directories, we need to have
access to the original parent directory where the prunable directories
are located. This commit defers the runtime directory setup to after
the config has been loaded so we can first prune the excess directories
and only then create the new runtime directory. Doing so, the validation
logic changes a bit as it's now fine to have unset runtime directory in
the config. SetupRuntimeDirectory knows how to handle this and will create
the temporary directory as done before.