cgroups: Check Processes before deleting cgroup
Now that gitaly processes will share cgroup directories between
processes, it's possible that some processes of the incoming Gitaly
process have started by the time the outgoing Gitaly tries to clean them
up.

In Cleanup, check if there are existing processes before deleting the
Cgroup. If there are, return a specific error that the caller can then
switch on. If they are probably leftover from an upgrade--in
which case we would want to allow them to continue.