cgroups: Cleanup cgroups during startup
Currently, cgroups Cleanup up is done upon shutdown. This can be
problematic however in an upgrade scenario when one Gitaly is shutting
down whilst another is coming online. There could be a race condition
where one is trying to delete cgroups while the other is trying to
create them.

To remove this complexity, move the Cleanup call to right before the
cgroups are setup.

Changelog: changed