cgroups: Add test for when Setup() called with existing cgroups dirs
Now that we have changed group paths from gitaly pid specific to to be
static, each gitaly process will be sharing cgroups directories during
upgrades. Hence we need to make sure that between upgrades, Setup() will
work and not error out if it finds existing cgroup directories.

Our cgroups library should handle this okay. To ensure this, add a test
that shows that when Setup is called when there are existing cgroup
directories, that it will not fail and instead will update any existing
cgroups with the new resource limits.