housekeeping: Improve heuristics for writing commit-graphs
Now that writing commit-graphs has been disentangled from repacking
objects we can iterate on the heuristics we use to determine whether we
need to write commit-graphs:

    - We stop writing commit-graphs in case there are no references.
      This wouldn't make any sense anyway given that we use only write
      commit-graphs for reachable commits, of which there aren't any in
      case there are no references.

    - We stop repacking objects based on whether the repository has
      bloom filters or not. This logic was previously mixed into the
      heuristics for `RepackObjects()`, but that has only be a historic
      artifact because that function handled both repacks and rewrites
      of commit-graphs. Instead, we only rewrite commit-graphs based on
      that information.

Ultimately, the end result should be that we repack objects less often,
but keep rewriting commit-graphs in the same circumstances as we did
before.

Add tests to verify this new behaviour and adjust existing tests for
repacking of objects.