housekeeping: Rewrite commit-graphs at a later point
Move writing of commit-graphs to happen after pruning of objects. This
is done so that we can start to use the information whether we pruned
any objects or not to decide whether to rewrite the commit-graph chain.

By moving the logic after packing refs we also have the benefit that we
have less loose refs to count in our heuristics.