housekeeping: Introduce configuration for WriteCommitGraph
Make the behaviour of whether we rewrite the commit-graph or not
controllable by callers and adjust callers accordingly. This new
configuration will be used to rewrite the commit-graph in more cases
than only missing bitmaps, as it is done right now.

While at it, improve the error messages to report what we're doing
instead of dictating policy like we do right now, where we ask the
caller to "please remove the commit-graph".