repository: Fix commit-graphs referencing pruned commits after PruneUnreachableObjects
When pruning objects we need to rewrite our commit-graphs so that they
don't reference any commits that have just been pruned. While this is
not typically an issue, git-fsck(1) will report any such inconsistencies
in the commit-graph. Furthermore, Git may return an error when being
asked to parse such a missing commit directly.

Fix `PruneUnreachableObjects()` to do so.

Changelog: fixed