repository: Refactor test to test GarbageCollect with locks
Refactor the test that verifies behaviour of the GarbageCollect RPC when
there are lockfiles around to match modern best practices.

Note that this refactoring surfaces an error when trying to garbage
collect a repo with a locked commit-graph. This is surfaces only now
because previously, we were reusing the same repository across all of
the tests. As a consequence, the commit-graph was getting generated by
the first subtest that called GarbageCollect without the lockfile having
been written yet. Subsequent calls thus wouldn't have to update it
anymore and succeeded even with the lockfile existing.

Now that we use a separate repository per test we get an error. This
error is in fact expected.