housekeeping: Repack object pools more aggressively
Object pools and normal repositories currently share the same heuristics
to decide whether to repack objects or not. While this strategy works,
it is not ideal for pool repositories because they may be reused across
many forks. It is thus desirable for us to keep them as efficient as
reasonably possible.

Tweak the heuristics to repack object pools more aggressively. Right now
this shouldn't make much of a difference: object pools are only updated
via `FetchIntoObjectPool()`, which knows to do a full repack when it has
fetched objects. As a result, no object pool should ever have more than
a single packfile in general.

This will make a difference though when we migrate object pools to not
do their own housekeeping anymore, but to use `OptimizeRepository()`
instead: we'll want to keep similar semantics as we have right now and
aggressively repack.

Note that we also need to fix some of our tests which used git-repack(1)
with the `-A` flag to bring repositories into the state we want to test
repository maintenance on. This will leave behind the old, superseded
packfile though and thus we end up with two packfiles even though the
intent of the test was to only test with a single packfile. This didn't
matter before where we required at least five packfiles before we pack
objects, but does now where the limit for object pools is at two packs.