objectpool: Improve error handling for FetchFromOrigin
Improve error handling in `FetchFromOrigin()` to provide proper
breakcrumbs so that it becomes easier to see where a specific error has
been raised. Furthermore, given that this is an internal package and not
an RPC's implementation, we shouldn't wrap errors with error codes but
instead leave this to the RPC implementations.