housekeeping: Speed up tests writing large packfiles via truncate(3p)
Some of our tests in the housekeeping package need to verify whether we
correctly compute the cutoff point when we want to do a full repack
based on the size of the largest packfile. We thus have to write such a
packfile into our temporary directory, which we're currently doing by
using some implementation that use `io.Reader`. This is inefficient, and
the tests spend a significant amount of time to just write these files.

Simplify the code and speed it up by just using truncate(3p) instead.
This speeds up the test from around 1s to 0.35s on my computer.