gittest: Work around corruption of commit-graphs caused by spooky commit
One of our test repositorise contains the "spooky commit" ba3343b (Weird
commit date, 292278994-08-17). This commit is rather far in the future,
and this is in fact corrupting commit-graphs:

    commit date for commit ba3343bc4fa403a8dfbfcab7fc1a8c29ee34bd69 in commit-graph is 15668040695 != 9223372036854775

We cannot really avoid this corruption for now, unfortunately, so we'll
just have to live with it for the time being until a fix to this bug
lands upstream.

This bug unfortunately does break tests which use this repository and
which compute and use commit-graphs. Right now there seemingly aren't
any, but we're about to add some that surface this breakage.

Work around this bug by deleting the reference to this broken commit and
prune the commit as required.