housekeeping: Convert function to check for pool repos to be typesafe
The `IsPoolPath()` function can be used to check whether a given path
belongs to a pool repository or not. It is not quite clear though what
kind of parameter the function expects: is it a relative path, or is it
the full path to the object pool? This is easy to get wrong and may have
fatal consequences.

Change the interface to instead accept a `repository.GitRepo` to clarify
this.