objectpool: Fix test for usage of delta islands
We have a test in our objectpool package that ought to verify whether we
set up delta islands correctly when fetching into an object pool. This
test is completely broken though: we don't check delta islands in the
pool, but instead in its primary pool member by accident.

Fix this bug and check delta islands in the pool itself. Furthermore,
stop repacking the pool objects: it destroys the purpose of the test as
we want to test exactly how objects are packed after fetching.

While at it, write a commit into the source repository before performing
the fetch so that there is actually data that has changed. This isn't
necessary right now, but is required as soon as wwe migrate maintenance
of pool repositories to use `OptimizeRepository()`.