housekeeping: Fix off-by-one in packfile count required to repack
The lower limit of packfiles required to exist in order to trigger a
repack has been picked as five packed. It is clearly documented as such,
and tests try to exercise this. There has been an off-by-one bug though,
which causes us to set the lower limit to six packfiles. And the tests
that intend to verify this behaviour have the same off-by-one bug.

Fix this off-by-one and make the heuristic work as advertised.

Changelog: fixed