datastore: Extend timeout to retrieve Postgres server's version
When booting up Praefect and connecting to Postgres, we check its
version so that we can be sure it's in fact running a supported version.
This check is using a timeout of a 100 milliseconds. While this probably
tends to work in most setups, this deadline can easily be busted when we
connect to a very busy server. And indeed, we frequently see this fail
on a particularly busy system, which is our CI.

Ideally, we'd just get rid of the timeout altogether to not run into
this scenario at all anymore. On the other hand it might be sensible to
fail on the startup of Praefect when we fail to connect to Postgres in a
reasonable amount of time so that it doesn't hang indefinitely. The real
question is what "reasonable amount of time" means though.

For now, let's pick a much more conservative timeout of five seconds.
This should hopefully be plenty of time for our CI, and still causes us
to fail comparatively fast on production systems in case anything goes
wrong.

Changelog: fixed