operations: Properly report unwrapped custom hook errors
Custom hook errors must be returned to the caller without modification
given that they may contain special prefixes that are parsed by the
Rails application. We have thus introduced the CustomHookError struct so
that we can unwrap that kind of error and then pass the error message to
the caller directly. But while we properly unwrap the error everywhere,
we don't in fact bubble up the contained error as expected.

Fix this and bubble up the unwrapped error's message. This allows us to
change error messages in those code paths to properly leave breadcrumbs
of the error trace.